Fix user ID lengths not being sufficient for MongoDB IDs

+++ Client/Trivia/Source/Codec/S2C/Response/Packets/Impl/IRegistrationResponse.cs
@@ -2,5 +2,5 @@ namespace Trivia.Codec.S2C.Response.Packets.Impl;
 
 public interface IRegistrationResponse : IProtocolResponse
 {
-    int UserId { get; init; }
+    long UserId { get; init; }
 }
